make of_a_kind return solutions always in order
to that effect, use pick instead of select. pick just returns the
remaining elements *right* of the picked element, not all of them.
The other change is that it's not necessary to pick the value in
advance: simply let it be an unbound variable, the first picked card
will then determine the value of the others.